Do not render "Number of colums" when create a sdg header card
In the admin widget cards form it is only treated as a header card when the
card has the header value set to true (card.header?). The problem we have in
the sdg_header_cards is that we don't set that field as true, but identify
it as header in another way.

It seems to make sense that from now on we set them as "header: true" but
that has implications in the code.

If we opt for this solution we would have to modify the code in the
following way:

Modify the "header_card?" method of the concern Admin::Widget::CardsActions
so that it would handle this case and set as "header = true" the sdg header
card we are going to create:

```
  def header_card?
    ++ params[:header_card].present? || params[:controller] == "sdg_management/header"
    -- params[:header_card].present?
  end
```

That would impicate that in the "self.header" method of the Widget::Card model
which is used to render the header cards in the admin section of the homepage
as follows:

```
  def self.header
    ++ where(header: true, cardable: nil)
    -- where(header: true)
  end
```

This change may be a bit difficult to follow, so instead of opting for this
solution and start treating the sdg header card as if it had the header field
set to true which could have other implications, we are going to improve the
condition that decides if a card is a header so that it will take into account
the case of sdg header cards. This way all the logic related to sdg header
cards still works the same.

Note that we are adding to the question:

`Card.header?`

the following condition:

`|| params[:controller] == "sdg_management/header"`

both to discern whether we show the "columns" field or not, and for the form
header messages, and for the submit texts. But we are not doing it for the line:

` <%= f.hidden_field :header, value: card.header? %>`

because as we have explained we don't want to treat it with the value header: true
to keep the logic as it is now.