Add a username slug to the user URL
This way it won't be possible to browse all user URLs by just going to
/users/1, /users/2, /users/3, ... and collect usernames, which might not
be desirable in some cases.

Note that we aren't doing the same thing in the direct messages
controller; in this controller, you can still use just the ID without
the username to access the URL. We think it isn't that risky because, in
order to send a direct message, you need to be a verified user.

Also note we could use the username as a URL parameter and just use
`@user = User.find_by!(id: id, username: username)`, but since usernames
might contain strange characters, this might lead to strange/ugly URLs.